ref: extract wizarding into its own module
